requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
joblib==1.2.0
jqdatasdk==1.8.11
jupyter_client==8.0.3
jupyter_core==5.2.0
kaggle==1.5.12
keras==2.11.0
kiwisolver==1.4.4
langchain==0.0.109
libclang==15.0.6.1
lightning-utilities==0.7.1
llama-index==0.4.28
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
opendatasets==0.1.22
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
ply==3.11
prompt-toolkit==3.0.37
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
Pygments==2.14.0
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytorch-lightning==1.9.3
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
thriftpy2==0.4.16
tiktoken==0.3.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.1
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
